# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Main Z11 api",
        "description": "This is our main Z11 api. It contains all method to handle Z11 project...",
        "termsOfService": "",
        "contact": {
            "email": "kienlv58@gmail.com"
        },
        "license": {
            "name": "Private License",
            "url": "URL to the license"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8000",
    "basePath": "/api/v1",
    "schemes": [
        "http"
    ],
    "paths": {
        "/answer/get_all/{take}/{skip}": {
            "get": {
                "tags": [
                    "8.Answer"
                ],
                "summary": "get all answer",
                "description": "return answer with take and skip",
                "operationId": "answer",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "take",
                        "in": "path",
                        "description": "take from ....",
                        "required": true,
                        "type": "integer",
                        "default": "all"
                    },
                    {
                        "name": "skip",
                        "in": "path",
                        "description": "skip from",
                        "required": true,
                        "type": "integer",
                        "default": "0"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid value"
                    }
                }
            }
        },
        "/answer/add": {
            "post": {
                "tags": [
                    "8.Answer"
                ],
                "summary": "add new answer",
                "description": "add new answer",
                "operationId": "answer_add",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "question_id",
                        "in": "formData",
                        "description": "question_id",
                        "required": true,
                        "schema": {
                            "required": [
                                "question_id"
                            ],
                            "type": "integer"
                        },
                        "type": "integer"
                    },
                    {
                        "name": "answer",
                        "in": "formData",
                        "description": "answer type json",
                        "required": true,
                        "schema": {
                            "required": [
                                "answer"
                            ],
                            "type": "string"
                        },
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "add succes"
                    },
                    "400": {
                        "description": "Invalid Value"
                    }
                }
            }
        },
        "/answer/edit": {
            "post": {
                "tags": [
                    "8.Answer"
                ],
                "summary": "edit answer",
                "description": "edit answer",
                "operationId": "question_edit",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "question_id",
                        "in": "formData",
                        "description": "question_id",
                        "required": true,
                        "schema": {
                            "required": [
                                "question_id"
                            ],
                            "type": "integer"
                        },
                        "type": "integer"
                    },
                    {
                        "name": "answer",
                        "in": "formData",
                        "description": "answer type json",
                        "required": true,
                        "schema": {
                            "required": [
                                "answer"
                            ],
                            "type": "string"
                        },
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "add succes"
                    },
                    "400": {
                        "description": "Invalid Value"
                    }
                }
            }
        },
        "/answer/delete": {
            "post": {
                "tags": [
                    "8.Answer"
                ],
                "summary": "delete answer ",
                "description": "delete with answer_id",
                "operationId": "answerdelete",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "uid",
                        "in": "formData",
                        "description": "uid delete",
                        "required": true,
                        "schema": {
                            "required": [
                                "uid"
                            ],
                            "type": "integer"
                        },
                        "type": "integer"
                    },
                    {
                        "name": "answer_item_id",
                        "in": "formData",
                        "description": "answer_item_id",
                        "required": true,
                        "schema": {
                            "required": [
                                "answer_item_id"
                            ],
                            "type": "integer"
                        },
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "delete succes"
                    },
                    "400": {
                        "description": "Invalid Value"
                    }
                }
            }
        },
...
        "/question/get_all/{take}/{skip}": {
            "get": {
                "tags": [
                    "7.Question"
                ],
                "summary": "get all question",
                "description": "return question with take and skip",
                "operationId": "question",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "take",
                        "in": "path",
                        "description": "take from ....",
                        "required": true,
                        "type": "integer",
                        "default": "all"
                    },
                    {
                        "name": "skip",
                        "in": "path",
                        "description": "skip from",
                        "required": true,
                        "type": "integer",
                        "default": "0"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid value"
                    }
                }
            }
        },
...
        "/question/edit": {
            "post": {
                "tags": [
                    "7.Question"
                ],
                "summary": "edit group_question",
                "description": "edit question",
                "operationId": "question_edit",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "question_id",
                        "in": "formData",
                        "description": "question_id",
                        "required": true,
                        "schema": {
                            "required": [
                                "question_id"
                            ],
                            "type": "integer"
                        },
                        "type": "integer"
                    },
                    {
                        "name": "sub_question_content",
                        "in": "formData",
                        "description": "sub_question_content",
                        "required": true,
                        "schema": {
                            "required": [
                                "sub_question_content"
                            ],
                            "type": "string"
                        },
                        "type": "string"
                    },
                    {
                        "name": "translate",
                        "in": "formData",
                        "description": "translate json",
                        "required": true,
                        "schema": {
                            "required": [
                                "category_code"
                            ],
                            "type": "string"
                        },
                        "type": "string"
                    },
                    {
                        "name": "explain_cost",
                        "in": "formData",
                        "description": "explain_cost",
                        "required": true,
                        "schema": {
                            "required": [
                                "explain_cost"
                            ],
                            "type": "string"
                        },
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "add succes"
                    },
                    "400": {
                        "description": "Invalid Value"
                    }
                }
            }
        },
...
    },
...
}